keep logic in Primitive.tsx
