+++ packages/react/src/primitives/actionBarMore/ActionBarMoreContent.tsx
@@ -9,7 +9,7 @@ import { DropdownMenu as DropdownMenuPrimitive } from "radix-ui";
 import {
   type WithRenderPropProps,
   withRenderProp,
-} from "../../utils/withRenderProp";
+} from "../../utils/Primitive";
 import { type ScopedProps, useDropdownMenuScope } from "./scope";
 
 const DropdownMenuContent = withRenderProp(DropdownMenuPrimitive.Content);
