+++ packages/react/src/primitives/actionBarMore/ActionBarMoreSeparator.tsx
@@ -5,7 +5,7 @@ import { DropdownMenu as DropdownMenuPrimitive } from "radix-ui";
 import {
   type WithRenderPropProps,
   withRenderProp,
-} from "../../utils/withRenderProp";
+} from "../../utils/Primitive";
 import { type ScopedProps, useDropdownMenuScope } from "./scope";
 
 const DropdownMenuSeparator = withRenderProp(DropdownMenuPrimitive.Separator);
